move to xpu.C, fix comments
